requirements.txt
Django==5.0.2
django_environ==0.11.2
django_rest_swagger==2.2.0
djangorestframework==3.14.0
drf_yasg==1.21.7
python-dotenv==1.0.1
psycopg[binary]==3.1.12
